Fixed tranfer logic error in Transfer SQL update query.
The transfer function can transfer between saving and current accounts.
However, due to the design of the transfer table that distinguishes accounts
based on their type rather from/to, the update query was wrong because
it always positive, meaning there's no way to tell which account is
transferring and which account is receiving.

I updated the query such that if transferring out of saving account, the
amount is negative in the transfer table. If into saving account, the
amount is positive.